+++ docs/changelog.md
@@ -8,6 +8,7 @@
 + Add "points alignment" function. (github issue #21)
 + Add "navigation toolbar position" option.
 + Add "auto remove empty links after deleted points" option. (github issue #23)
++ Move background option into project file.
 
 **Development**: